backend/requirements.txt
Django>=3.0,<4.0
psycopg2==2.9.6
asgiref==3.6.0
djangorestframework==3.14.0
pytz==2023.3
sqlparse==0.4.3
